self_improving_assistant/requirements.txt
fastapi
uvicorn
pyyaml
pydantic
tiktoken
openai
requests
beautifulsoup4
lxml
duckduckgo-search
rank-bm25
